[sourcemaps] Improve ignore-listing performance
This improves performance of finding the relevant
section in Index Source Maps by using binary search
(O(N) -> O(log N)).

This allows us to improve the performance of the
fast path for checking if a source is ignore-listed in
Index Source Maps:
Instead of checking if every section in Index Source
Maps ignores everything, we just find the relevant
section (now O(log N) instead of O(N)) and check if
that section ignores everything.